tests/utils: rescoped ResultsFileDumper
* Added as well an ExceptionAggregator abstraction.
